Rtrepiana

+++ aon-db-up2date/src/main/java/net/aonsolutions/db/up2date/fiscal/RefreshMod131Result.java
@@ -81,7 +81,7 @@ public class RefreshMod131Result implements Update {
 							if (result != null) {
 								dslContext.update(FS_MODEL)
 									.set(FS_MODEL.RESULT,result)
-									.set(FS_MODEL.DECLARATION_TYPE, (declarationType == null? null : declarationType.value()))
+									.set(FS_MODEL.DECLARATION_TYPE, (declarationType == null? null : (Byte)declarationType.value()))
 									.where(FS_MODEL.ID.eq(id))
 									.execute();
 							}
